backend/requirements.txt
fastapi
uvicorn
pandas
langchain-huggingface
langchain-community
langchain-text-splitters
qdrant-client
langchain-qdrant
pydantic
supabase
sentence-transformers
httpx
redis
python-dotenv